test-LTE-GPS/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
idna==3.11
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
pillow==12.0.0
pyhackrf2==1.0.3
pyparsing==3.2.5
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
requests==2.32.5
scipy==1.16.3
six==1.17.0
urllib3==2.5.0
zmq==0.0.0
